+++ src/xSocket.cpp
@@ -3,7 +3,7 @@
 #include "xLog.h"
 
 
-xSocket::xSocket(xEventLoop *loop,std::string ip,int32_t port)
+xSocket::xSocket(xEventLoop *loop,std::string ip, int16_t port)
 :loop(loop),
  ip(ip),
  port(port)
@@ -33,7 +33,7 @@ int  xSocket::getListenFd()
 }
 
 
-bool xSocket::getpeerName(int32_t fd,std::string *ip,int32_t &port)
+bool xSocket::getpeerName(int32_t fd,std::string *ip, int16_t &port)
 {
 	struct sockaddr_in sa;
 	socklen_t len = sizeof(sa);
@@ -57,7 +57,7 @@ int  xSocket::createNonBloackSocket()
 	return socket(AF_INET, SOCK_STREAM, 0);
 }
 
-int xSocket::connect(int sockfd,std::string ip,int port)
+int xSocket::connect(int sockfd,std::string ip, int16_t port)
 {
 	struct sockaddr_in sin;
 	memset(&sin, 0, sizeof(sin));
